Separated classes/interfaces to their own files
