+++ AnnoDesigner/CommandLine/Arguments/AdminRestartArgs.cs
@@ -1,0 +1,10 @@
+﻿using CommandLine;
+
+namespace AnnoDesigner.CommandLine.Arguments
+{
+    [Verb("askAdmin", Hidden = true)]
+    public class AdminRestartArgs : IProgramArgs
+    {
+        public const string Arguments = "askAdmin";
+    }
+}
