+++ AnnoDesigner/CommandLine/Arguments/ExportArgs.cs
@@ -1,25 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using CommandLine;
-using CommandLine.Text;
+﻿using CommandLine;
 
-namespace AnnoDesigner
+namespace AnnoDesigner.CommandLine.Arguments
 {
-    public interface IProgramArgs { }
-
-    [Verb("askAdmin", Hidden = true)]
-    public class AdminRestartArgs : IProgramArgs
-    {
-        public const string Arguments = "askAdmin";
-    }
-
-    [Verb("open", HelpText = "Opens layout after AnnoDesigner starts")]
-    public class OpenArgs : IProgramArgs
-    {
-        [Value(0, HelpText = "Input AD file", Required = true)]
-        public string Filename { get; set; }
-    }
-
     [Verb("export", HelpText = "Exports layout to image and exits")]
     public class ExportArgs : IProgramArgs
     {
@@ -61,31 +43,5 @@ namespace AnnoDesigner
 
         [Option("renderVersion")]
         public bool RenderVersion { get; set; }
-    }
-
-    public class HelpException : Exception
-    {
-        public HelpText HelpText { get; }
-
-        public HelpException(HelpText helpText)
-        {
-            HelpText = helpText;
-        }
-    }
-
-    public static class ArgumentParser
-    {
-        public static IProgramArgs Parse(IEnumerable<string> arguments)
-        {
-            var parsed = Parser.Default.ParseArguments<AdminRestartArgs, OpenArgs, ExportArgs>(arguments);
-            return parsed.MapResult<AdminRestartArgs, OpenArgs, ExportArgs, IProgramArgs>(x => x, x => x, x => x, errors =>
-            {
-                if (errors.IsHelp() || errors.IsVersion())
-                {
-                    throw new HelpException(HelpText.AutoBuild(parsed));
-                }
-                return null;
-            });
-        }
     }
 }
