+++ AnnoDesigner/CommandLine/Arguments/IProgramArgs.cs
@@ -1,0 +1,4 @@
+﻿namespace AnnoDesigner.CommandLine.Arguments
+{
+    public interface IProgramArgs { }
+}
