+++ AnnoDesigner/CommandLine/Arguments/OpenArgs.cs
@@ -1,0 +1,11 @@
+﻿using CommandLine;
+
+namespace AnnoDesigner.CommandLine.Arguments
+{
+    [Verb("open", HelpText = "Opens layout after AnnoDesigner starts")]
+    public class OpenArgs : IProgramArgs
+    {
+        [Value(0, HelpText = "Input AD file", Required = true)]
+        public string Filename { get; set; }
+    }
+}
